# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Agenda Escolar
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Subject {
  id: number;
  name: string;
  teacher?: string;
  /** Hex color for the subject */
  color: string;
  createdAt: string;
}

export interface CreateSubjectBody {
  name: string;
  teacher?: string;
  color: string;
}

export interface ScheduleEntry {
  id: number;
  subjectId: number;
  subjectName: string;
  subjectColor: string;
  /** 0=Sunday, 1=Monday, ..., 6=Saturday */
  dayOfWeek: number;
  /** HH:MM format */
  startTime: string;
  /** HH:MM format */
  endTime: string;
  room?: string;
}

export interface CreateScheduleEntryBody {
  subjectId: number;
  dayOfWeek: number;
  startTime: string;
  endTime: string;
  room?: string;
}

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority];

export const TaskPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Task {
  id: number;
  title: string;
  description?: string;
  subjectId?: number;
  subjectName?: string;
  subjectColor?: string;
  dueDate?: string;
  completed: boolean;
  priority: TaskPriority;
  createdAt: string;
}

export type CreateTaskBodyPriority =
  (typeof CreateTaskBodyPriority)[keyof typeof CreateTaskBodyPriority];

export const CreateTaskBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface CreateTaskBody {
  title: string;
  description?: string;
  subjectId?: number;
  dueDate?: string;
  priority: CreateTaskBodyPriority;
}

export type UpdateTaskBodyPriority =
  (typeof UpdateTaskBodyPriority)[keyof typeof UpdateTaskBodyPriority];

export const UpdateTaskBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface UpdateTaskBody {
  title?: string;
  description?: string;
  subjectId?: number;
  dueDate?: string;
  completed?: boolean;
  priority?: UpdateTaskBodyPriority;
}

export type EventType = (typeof EventType)[keyof typeof EventType];

export const EventType = {
  exam: "exam",
  presentation: "presentation",
  fieldtrip: "fieldtrip",
  holiday: "holiday",
  other: "other",
} as const;

export interface Event {
  id: number;
  title: string;
  description?: string;
  subjectId?: number;
  subjectName?: string;
  subjectColor?: string;
  date: string;
  type: EventType;
  createdAt: string;
}

export type CreateEventBodyType =
  (typeof CreateEventBodyType)[keyof typeof CreateEventBodyType];

export const CreateEventBodyType = {
  exam: "exam",
  presentation: "presentation",
  fieldtrip: "fieldtrip",
  holiday: "holiday",
  other: "other",
} as const;

export interface CreateEventBody {
  title: string;
  description?: string;
  subjectId?: number;
  date: string;
  type: CreateEventBodyType;
}

export interface Note {
  id: number;
  title: string;
  content: string;
  subjectId?: number;
  subjectName?: string;
  subjectColor?: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateNoteBody {
  title: string;
  content: string;
  subjectId?: number;
}

export interface DashboardSummary {
  pendingTasksCount: number;
  completedTasksCount: number;
  upcomingEventsCount: number;
  subjectsCount: number;
  upcomingTasks: Task[];
  upcomingEvents: Event[];
  todaySchedule: ScheduleEntry[];
}

export type ListTasksParams = {
  /**
   * Filter by status
   */
  status?: ListTasksStatus;
};

export type ListTasksStatus =
  (typeof ListTasksStatus)[keyof typeof ListTasksStatus];

export const ListTasksStatus = {
  pending: "pending",
  completed: "completed",
  all: "all",
} as const;

export type ListEventsParams = {
  /**
   * Filter for upcoming events only
   */
  upcoming?: boolean;
};