# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Agenda Escolar
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: subjects
    description: School subjects
  - name: schedule
    description: Weekly schedule
  - name: tasks
    description: Homework and tasks
  - name: events
    description: School events and exams
  - name: notes
    description: General notes
  - name: dashboard
    description: Dashboard summary

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /subjects:
    get:
      operationId: listSubjects
      tags: [subjects]
      summary: List all subjects
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
    post:
      operationId: createSubject
      tags: [subjects]
      summary: Create a subject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubjectBody"
      responses:
        "201":
          description: Subject created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"

  /subjects/{id}:
    put:
      operationId: updateSubject
      tags: [subjects]
      summary: Update a subject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubjectBody"
      responses:
        "200":
          description: Subject updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    delete:
      operationId: deleteSubject
      tags: [subjects]
      summary: Delete a subject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Subject deleted

  /schedule:
    get:
      operationId: listSchedule
      tags: [schedule]
      summary: List all schedule entries
      responses:
        "200":
          description: List of schedule entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleEntry"
    post:
      operationId: createScheduleEntry
      tags: [schedule]
      summary: Create a schedule entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScheduleEntryBody"
      responses:
        "201":
          description: Schedule entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleEntry"

  /schedule/{id}:
    put:
      operationId: updateScheduleEntry
      tags: [schedule]
      summary: Update a schedule entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScheduleEntryBody"
      responses:
        "200":
          description: Schedule entry updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleEntry"
    delete:
      operationId: deleteScheduleEntry
      tags: [schedule]
      summary: Delete a schedule entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Schedule entry deleted

  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, completed, all]
          description: Filter by status
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/{id}:
    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Task deleted

  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List all events
      parameters:
        - name: upcoming
          in: query
          schema:
            type: boolean
          description: Filter for upcoming events only
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"

  /events/{id}:
    put:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "200":
          description: Event updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Event deleted

  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List all notes
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteBody"
      responses:
        "201":
          description: Note created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"

  /notes/{id}:
    put:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteBody"
      responses:
        "200":
          description: Note updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Note deleted

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary
      responses:
        "200":
          description: Dashboard summary with counts and upcoming items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Subject:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        teacher:
          type: string
        color:
          type: string
          description: Hex color for the subject
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - color
        - createdAt

    CreateSubjectBody:
      type: object
      properties:
        name:
          type: string
        teacher:
          type: string
        color:
          type: string
      required:
        - name
        - color

    ScheduleEntry:
      type: object
      properties:
        id:
          type: integer
        subjectId:
          type: integer
        subjectName:
          type: string
        subjectColor:
          type: string
        dayOfWeek:
          type: integer
          description: "0=Sunday, 1=Monday, ..., 6=Saturday"
        startTime:
          type: string
          description: "HH:MM format"
        endTime:
          type: string
          description: "HH:MM format"
        room:
          type: string
      required:
        - id
        - subjectId
        - subjectName
        - subjectColor
        - dayOfWeek
        - startTime
        - endTime

    CreateScheduleEntryBody:
      type: object
      properties:
        subjectId:
          type: integer
        dayOfWeek:
          type: integer
        startTime:
          type: string
        endTime:
          type: string
        room:
          type: string
      required:
        - subjectId
        - dayOfWeek
        - startTime
        - endTime

    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        subjectId:
          type: integer
        subjectName:
          type: string
        subjectColor:
          type: string
        dueDate:
          type: string
          format: date
        completed:
          type: boolean
        priority:
          type: string
          enum: [low, medium, high]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - completed
        - priority
        - createdAt

    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        subjectId:
          type: integer
        dueDate:
          type: string
          format: date
        priority:
          type: string
          enum: [low, medium, high]
      required:
        - title
        - priority

    UpdateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        subjectId:
          type: integer
        dueDate:
          type: string
          format: date
        completed:
          type: boolean
        priority:
          type: string
          enum: [low, medium, high]

    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        subjectId:
          type: integer
        subjectName:
          type: string
        subjectColor:
          type: string
        date:
          type: string
          format: date
        type:
          type: string
          enum: [exam, presentation, fieldtrip, holiday, other]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - date
        - type
        - createdAt

    CreateEventBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        subjectId:
          type: integer
        date:
          type: string
          format: date
        type:
          type: string
          enum: [exam, presentation, fieldtrip, holiday, other]
      required:
        - title
        - date
        - type

    Note:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        subjectId:
          type: integer
        subjectName:
          type: string
        subjectColor:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - content
        - createdAt
        - updatedAt

    CreateNoteBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        subjectId:
          type: integer
      required:
        - title
        - content

    DashboardSummary:
      type: object
      properties:
        pendingTasksCount:
          type: integer
        completedTasksCount:
          type: integer
        upcomingEventsCount:
          type: integer
        subjectsCount:
          type: integer
        upcomingTasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
        upcomingEvents:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        todaySchedule:
          type: array
          items:
            $ref: "#/components/schemas/ScheduleEntry"
      required:
        - pendingTasksCount
        - completedTasksCount
        - upcomingEventsCount
        - subjectsCount
        - upcomingTasks
        - upcomingEvents
        - todaySchedule