# lib/api-zod/src/generated/types/createScheduleEntryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Agenda Escolar
 * OpenAPI spec version: 0.1.0
 */

export interface CreateScheduleEntryBody {
  subjectId: number;
  dayOfWeek: number;
  startTime: string;
  endTime: string;
  room?: string;
}