# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Agenda Escolar
 * OpenAPI spec version: 0.1.0
 */
import type { Event } from "./event";
import type { ScheduleEntry } from "./scheduleEntry";
import type { Task } from "./task";

export interface DashboardSummary {
  pendingTasksCount: number;
  completedTasksCount: number;
  upcomingEventsCount: number;
  subjectsCount: number;
  upcomingTasks: Task[];
  upcomingEvents: Event[];
  todaySchedule: ScheduleEntry[];
}