# lib/api-zod/src/generated/types/scheduleEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Agenda Escolar
 * OpenAPI spec version: 0.1.0
 */

export interface ScheduleEntry {
  id: number;
  subjectId: number;
  subjectName: string;
  subjectColor: string;
  /** 0=Sunday, 1=Monday, ..., 6=Saturday */
  dayOfWeek: number;
  /** HH:MM format */
  startTime: string;
  /** HH:MM format */
  endTime: string;
  room?: string;
}