# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Agenda Escolar
 * OpenAPI spec version: 0.1.0
 */
import type { TaskPriority } from "./taskPriority";

export interface Task {
  id: number;
  title: string;
  description?: string;
  subjectId?: number;
  subjectName?: string;
  subjectColor?: string;
  dueDate?: Date;
  completed: boolean;
  priority: TaskPriority;
  createdAt: Date;
}